src/requirements.txt
Django==4.1.3
djangorestframework==3.14.0
django-bootstrap5==22.2
Pillow==10.0.0
tzdata==2023.3
biopython==1.80
pandas==1.5.2
numpy==1.23.5
scipy==1.9.3
matplotlib==3.7.2
openpyxl==3.1.2
